Remove Configuration class, move everything to Equilibrium
